Remove deploy-ssh container
It's not been being maintained. The version of SSH & accompanying utils should be fine on the GitHub build agent. If it is, then that is better as the GitHub agents will be being patched, but that docker image isn't.

This was originally carried over from the GitLab pipeline, and probably isn't necessary here.